Add start.sh (read description to learn more)
This file is a script that accompanies an herbstluftwm desktop
environment. It will automatically fill in the 9 tags with the content I
normally have on each one. Over the years I've grown accustomed to
certain tags having certain windows on them. For instance:

1: Web Browser
2: Code Editor (launched from tag 0's toolbox)
3: 4 terminals taking equal space
4: Database Viewer (launched from tag 0)
5: telegram (chatting), mailspring (email), and slack (slack)
6: Postman / Insomnia (REST client)
7: Temp / Notes
8: Temp
9: Temp
0: Toolbox, Music, Misc.

This script can be customized to use your tag names and programs